Added tests, fixed bugs.

+++ PNGAnalyzer/PNGAnalyzer/PNGAnalyzer/Chunks/iCCP.cs
@@ -26,7 +26,7 @@ namespace PNGAnalyzer
 
         private void ParseData(byte[] data)
         {
-            int nullSeparatorIndex = Array.IndexOf(data, 0);
+            int nullSeparatorIndex = Array.IndexOf(data, (byte) 0);
             ProfileName = Encoding.GetEncoding("iso-8859-1").GetString(data, 0, nullSeparatorIndex);
             CompressionMethod = data[nullSeparatorIndex + 1];
             CompressedProfile = data.Skip(nullSeparatorIndex + 2).ToArray();
